requirements.txt
boto3==1.17.28
flask==1.1.2
gunicorn==20.0.4
h5py==2.10.0
imageio==2.9.0
jedi==0.17.0
Jinja2==2.11.3
json5==0.9.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
notebook==6.3.0
numpy==1.18.0
oauthlib==3.1.0
pandas==1.1.3
pandocfilters==1.4.3
pickleshare==0.7.5
Pillow==8.0.0
pip==21.0.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyreadline==2.1
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
requests==2.25.1
requests-oauthlib==1.3.0
scipy==1.4.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
tweepy==3.10.0
wheel==0.36.2
